fix: fix cost

+++ src/main/java/com/event/wear/platform/publication/interfaces.rest/resources/PublicationResource.java
@@ -1,4 +1,4 @@
 package com.event.wear.platform.publication.interfaces.rest.resources;
 
-public record PublicationResource (Integer id, String image, String title, String description, Integer price, Integer rating){
+public record PublicationResource (Long id, String image, String title, String description, Integer price, Integer rating){
 }
